refactor: move prisma dir back and fix file imports
