+++ prisma/schema.prisma
@@ -1,0 +1,110 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+// Use Prisma Migrate if you change the schema again
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+model TriviaGame {
+  id           Int         @default(autoincrement()) @id
+  createdAt    DateTime    @default(now()) @map(name: "created_at")
+  playedAt     DateTime?   @map(name: "played_at")
+  hostId       Int?
+  roundAmount  Int
+  questions    Question[]
+  //need to add Teams?
+}
+
+model Question {
+  id             Int        @default(autoincrement()) @id
+  trivia         TriviaGame   @relation(fields: [triviaId], references: [id])
+  triviaId       Int
+  roundNum       Int
+  questionNum    Int?
+  content        String?
+  type           String?
+  correctAnswer  String?
+  answers        Answer[]
+}
+
+model Answer {
+  id            Int         @default(autoincrement()) @id
+  question      Question    @relation(fields: [questionId], references: [id])
+  questionId    Int
+  // teamId        Int
+  teamName      String
+  content       String?
+  isCorrect     Boolean     @default(false)
+  submittedAt   DateTime    @default(now())
+}
+
+model Team {
+  id            Int         @default(autoincrement()) @id
+  teamName      String?
+  members       User[]
+  @@map(name: "teams")
+}
+
+
+model User {
+  id            Int         @default(autoincrement()) @id
+  name          String?
+  displayName   String?
+  email         String?     @unique
+  emailVerified DateTime? @map(name: "email_verified")
+  image         String?
+  createdAt     DateTime?    @default(now()) @map(name: "created_at")
+  updatedAt     DateTime?   @updatedAt @map(name: "updated_at")
+  teams         Team[]
+  @@map(name: "users")
+}
+
+model Account {
+  id                 Int       @default(autoincrement()) @id
+  compoundId         String    @unique @map(name: "compound_id")
+  userId             Int       @map(name: "user_id")
+  providerType       String    @map(name: "provider_type")
+  providerId         String    @map(name: "provider_id")
+  providerAccountId  String    @map(name: "provider_account_id")
+  refreshToken       String?   @map(name: "refresh_token")
+  accessToken        String?   @map(name: "access_token")
+  accessTokenExpires DateTime? @map(name: "access_token_expires")
+  createdAt          DateTime  @default(now()) @map(name: "created_at")
+  updatedAt          DateTime  @default(now()) @map(name: "updated_at")
+
+  @@index([providerAccountId], name: "providerAccountId")
+  @@index([providerId], name: "providerId")
+  @@index([userId], name: "userId")
+
+  @@map(name: "accounts")
+}
+
+model Session {
+  id           Int      @default(autoincrement()) @id
+  userId       Int      @map(name: "user_id")
+  expires      DateTime
+  sessionToken String   @unique @map(name: "session_token")
+  accessToken  String   @unique @map(name: "access_token")
+  createdAt    DateTime @default(now()) @map(name: "created_at")
+  updatedAt    DateTime @default(now()) @map(name: "updated_at")
+
+  @@map(name: "sessions")
+}
+
+model VerificationRequest {
+  id         Int      @default(autoincrement()) @id
+  identifier String
+  token      String   @unique
+  expires    DateTime
+  createdAt  DateTime  @default(now()) @map(name: "created_at")
+  updatedAt  DateTime  @default(now()) @map(name: "updated_at")
+
+  @@map(name: "verification_requests")
+}
